huge file system update also havent tested this all in a while but I will do that tomorrow
