+++ package.json
@@ -45,10 +45,10 @@
 	},
 	"dependencies": {
 		"@discordjs/builders": "^0.12.0",
-		"@types/node-fetch": "^2.6.1",
 		"discord-api-types": "^0.29.0",
 		"discord.js": "^13.6.0",
 		"dotenv": "^16.0.0",
+		"node-fetch": "^3.2.3",
 		"prettier": "^2.6.0",
 		"winston": "^3.6.0"
 	}
