+++ src/util/countryCommand/generalCountryCommandUtil.ts
@@ -1,0 +1,9 @@
+import { CommandInteraction } from "discord.js";
+
+export function countryUndefinedReply(interaction: CommandInteraction) {
+    interaction.reply({
+        content: "I've never heard of that country... Next time pick one from the list, okay?",
+        ephemeral: true,
+    });
+}
+
